# opt:(ast) pass `skipnumber` flag to avoid decoding numbers