--- a/README.md
+++ b/README.md
@@ -6,7 +6,7 @@ A blazingly fast JSON serializing &amp; deserializing library, accelerated by JI
 
 ## Requirement
 - Go 1.16~1.20
-- Linux/MacOS/Windows(需要go1.17以上)
+- Linux/MacOS/Windows(need go1.17 above)
 - Amd64 ARCH
 
 ## Features
@@ -76,6 +76,8 @@ BenchmarkSetOne_Jsoniter-16                            79475 ns/op         163.8
 BenchmarkSetOne_Parallel_Sonic-16                      850.9 ns/op       15305.31 MB/s        1584 B/op         17 allocs/op
 BenchmarkSetOne_Parallel_Sjson-16                      18194 ns/op         715.77 MB/s       52247 B/op          9 allocs/op
 BenchmarkSetOne_Parallel_Jsoniter-16                   33560 ns/op         388.05 MB/s       45892 B/op        964 allocs/op
+BenchmarkLoadNode/LoadAll()-16                          4918 ns/op        2647.98 MB/s        7175 B/op         25 allocs/op
+BenchmarkLoadNode/Interface()-16                        9417 ns/op        1382.89 MB/s       15425 B/op         88 allocs/op
 ```
 - [Small](https://github.com/bytedance/sonic/blob/main/testdata/small.go) (400B, 11 keys, 3 layers)
 ![small benchmarks](./docs/imgs/bench-small.png)