--- a/ast/parser.go
+++ b/ast/parser.go
@@ -195,7 +195,7 @@ func (self *Parser) decodeObject(ret *linkedPairs) (Node, types.ParsingError) {
         var err types.ParsingError
 
         /* decode the key */
-        if njs = self.decodeValue(); njs.Vt != types.V_STRING {
+        if njs = self.decodeValue(true); njs.Vt != types.V_STRING {
             return Node{}, types.ERR_INVALID_CHAR
         }
 
@@ -287,7 +287,7 @@ func (self *Parser) Pos() int {
 }
 
 func (self *Parser) Parse() (Node, types.ParsingError) {
-    switch val := self.decodeValue(); val.Vt {
+    switch val := self.decodeValue(true); val.Vt {
         case types.V_EOF     : return Node{}, types.ERR_EOF
         case types.V_NULL    : return nullNode, 0
         case types.V_TRUE    : return trueNode, 0
@@ -340,7 +340,7 @@ func (self *Parser) searchKey(match string) types.ParsingError {
     for {
 
         /* decode the key */
-        if njs = self.decodeValue(); njs.Vt != types.V_STRING {
+        if njs = self.decodeValue(true); njs.Vt != types.V_STRING {
             return types.ERR_INVALID_CHAR
         }
 
@@ -521,7 +521,7 @@ func (self *Node) skipNextPair() (*Pair) {
     var err types.ParsingError
 
     /* decode the key */
-    if njs = parser.decodeValue(); njs.Vt != types.V_STRING {
+    if njs = parser.decodeValue(true); njs.Vt != types.V_STRING {
         return &Pair{"", *newSyntaxError(parser.syntaxError(types.ERR_INVALID_CHAR))}
     }