## Benchmark
```
name                                     old time/op    new time/op    delta
LoadNode_Parallel/LoadAll()-16             6.08µs ± 4%    5.76µs ±26%     ~     (p=0.203 n=8+10)
LoadNode_Parallel/Load()-16                1.80µs ± 3%    2.02µs ±22%     ~     (p=0.055 n=10+10)
LoadNode/Load()-16                         5.54µs ± 5%    5.16µs ± 2%   -6.97%  (p=0.000 n=10+10)
LoadNode_Parallel/InterfaceUseNode()-16    2.91µs ± 5%    2.68µs ±16%   -7.80%  (p=0.003 n=10+9)
LoadNode/InterfaceUseNode()-16             6.77µs ± 4%    6.22µs ± 4%   -8.18%  (p=0.000 n=10+9)
LoadNode_Parallel/Interface()-16           12.1µs ± 2%    10.3µs ± 4%  -14.36%  (p=0.000 n=6+10)
LoadNode/LoadAll()-16                      13.6µs ± 8%    11.5µs ±10%  -15.28%  (p=0.000 n=10+10)
LoadNode/Interface()-16                    26.7µs ±18%    17.7µs ± 0%  -33.59%  (p=0.000 n=10+8)

name                                     old speed      new speed      delta
LoadNode_Parallel/LoadAll()-16           2.11GB/s ±11%  2.29GB/s ±21%     ~     (p=0.133 n=9+10)
LoadNode_Parallel/Load()-16              7.24GB/s ± 3%  6.56GB/s ±19%     ~     (p=0.063 n=10+10)
LoadNode/Load()-16                       2.35GB/s ± 5%  2.53GB/s ± 2%   +7.46%  (p=0.000 n=10+10)
LoadNode/InterfaceUseNode()-16           1.92GB/s ± 4%  2.09GB/s ± 4%   +8.89%  (p=0.000 n=10+9)
LoadNode_Parallel/InterfaceUseNode()-16  4.48GB/s ± 5%  4.87GB/s ±14%   +8.89%  (p=0.003 n=10+9)
LoadNode_Parallel/Interface()-16         1.08GB/s ± 2%  1.26GB/s ± 4%  +16.84%  (p=0.000 n=6+10)
LoadNode/LoadAll()-16                     960MB/s ± 8%  1135MB/s ±10%  +18.21%  (p=0.000 n=10+10)
LoadNode/Interface()-16                   493MB/s ±16%   734MB/s ± 0%  +49.02%  (p=0.000 n=10+8)

name                                     old alloc/op   new alloc/op   delta
LoadNode/Load()-16                         1.06kB ± 0%    1.09kB ± 0%   +2.95%  (p=0.000 n=9+10)
LoadNode_Parallel/Load()-16                1.20kB ± 1%    1.24kB ± 1%   +2.75%  (p=0.000 n=10+10)
LoadNode_Parallel/InterfaceUseNode()-16    2.99kB ± 1%    3.03kB ± 3%   +1.31%  (p=0.004 n=10+10)
LoadNode/InterfaceUseNode()-16             2.62kB ± 0%    2.65kB ± 0%   +1.31%  (p=0.000 n=10+10)
LoadNode/Interface()-16                    13.3kB ± 0%    13.3kB ± 0%   +0.25%  (p=0.000 n=10+10)
LoadNode/LoadAll()-16                      6.30kB ± 0%    6.31kB ± 0%   +0.06%  (p=0.045 n=10+9)
LoadNode_Parallel/Interface()-16           15.0kB ± 2%    15.1kB ± 1%     ~     (p=0.243 n=9+10)
LoadNode_Parallel/LoadAll()-16             7.08kB ± 2%    7.10kB ± 2%     ~     (p=0.684 n=10+10)

name                                     old allocs/op  new allocs/op  delta
LoadNode/Interface()-16                      88.0 ± 0%      88.0 ± 0%     ~     (all equal)
LoadNode/LoadAll()-16                        25.0 ± 0%      25.0 ± 0%     ~     (all equal)
LoadNode/InterfaceUseNode()-16               7.00 ± 0%      7.00 ± 0%     ~     (all equal)
LoadNode/Load()-16                           5.00 ± 0%      5.00 ± 0%     ~     (all equal)
LoadNode_Parallel/Interface()-16             88.0 ± 0%      88.0 ± 0%     ~     (all equal)
LoadNode_Parallel/LoadAll()-16               25.0 ± 0%      25.0 ± 0%     ~     (all equal)
LoadNode_Parallel/InterfaceUseNode()-16      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
LoadNode_Parallel/Load()-16                  5.00 ± 0%      5.00 ± 0%     ~     (all equal)
```